+++ core/src/main/resources/irace/runner.R
@@ -1,5 +1,0 @@
-install.packages("irace")
-library("irace")
-scenario <- readScenario(filename = "scenario.txt", scenario = defaultScenario())
-checkIraceScenario(scenario = scenario)
-irace.main(scenario = scenario)